pyproject.toml
[project]
name = "whatsapp-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "datetime>=6.0",
    "google-genai>=1.61.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "selenium>=4.40.0",
]
